+++ bootloader/src/pyi_utils_win32.c
@@ -25,6 +25,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sddl.h> /* ConvertStringSecurityDescriptorToSecurityDescriptorW */
 
 /* PyInstaller headers. */
 #include "pyi_utils.h"
@@ -530,6 +531,212 @@ pyi_utils_create_child(PYI_CONTEXT *pyi_ctx)
     FATAL_WINERROR("CreateProcessW", "Failed to create child process!\n");
     return -1;
 }
+
+
+/**********************************************************************\
+ *             Security descriptor for temporary directory            *
+\**********************************************************************/
+/* Retrieve the SID for the specified token information class from the current process.
+ *
+ * At the moment, TokenUser and TokenAppContainerSid are supported.
+ *
+ * Used in a compatibility work-around for wine, which at the time of writing
+ * (version 5.0.2) does not properly support SID S-1-3-4 (directory owner),
+ * and therefore user's actual SID must be used instead.
+ *
+ * Returns a copy of SID string on success, NULL on failure (or if the SID is unavailable
+ * or zero-length). The returned string must be freed using LocalFree().
+ */
+static wchar_t *
+_pyi_win32_get_sid(TOKEN_INFORMATION_CLASS token_information_class)
+{
+    HANDLE process_token = INVALID_HANDLE_VALUE;
+    DWORD token_info_size = 0;
+    void *token_info = NULL;
+    wchar_t *sid = NULL;
+
+    /* Get access token for the calling process */
+    if (!OpenProcessToken(GetCurrentProcess(), TOKEN_QUERY, &process_token)) {
+        goto cleanup;
+    }
+
+    /* Query buffer size and allocate buffer */
+    if (!GetTokenInformation(process_token, token_information_class, NULL, 0, &token_info_size)) {
+        if (GetLastError() != ERROR_INSUFFICIENT_BUFFER) {
+            goto cleanup;
+        }
+    }
+    if (token_info_size == 0) {
+        /* As per MSDN, in the Microsoft implementation, if number or size is zero, calloc
+         * returns a pointer to an allocated block of non-zero size. An attempt to read or
+         * write through the returned pointer leads to undefined behavior. */
+        goto cleanup;
+    }
+    token_info = calloc(1, token_info_size);
+    if (!token_info) {
+        goto cleanup;
+    }
+
+    /* Get token information */
+    if (!GetTokenInformation(process_token, token_information_class, token_info, token_info_size, &token_info_size)) {
+        goto cleanup;
+    }
+
+    /* Convert SID to string */
+    switch (token_information_class)
+    {
+        case TokenUser: {
+            PTOKEN_USER user_info = (PTOKEN_USER)token_info;
+            ConvertSidToStringSidW(user_info->User.Sid, &sid);
+            break;
+        }
+        case TokenAppContainerSid: {
+            PTOKEN_APPCONTAINER_INFORMATION app_container_info = (PTOKEN_APPCONTAINER_INFORMATION)token_info;
+            ConvertSidToStringSidW(app_container_info->TokenAppContainer, &sid);
+            break;
+        }
+        default: {
+            /* Unsupoorted token information class */
+            break;
+        }
+    }
+
+    // Cleanup
+cleanup:
+    free(token_info);
+    if (process_token != INVALID_HANDLE_VALUE) {
+        CloseHandle(process_token);
+    }
+
+    return sid;
+}
+
+/* Initialize security descriptor applied to application's temporary directory and its
+ * sub-directories.
+ */
+SECURITY_ATTRIBUTES *
+pyi_win32_initialize_security_descriptor()
+{
+    SECURITY_ATTRIBUTES *security_attr;
+    LPVOID lpSecurityDescriptor;
+    wchar_t *user_sid;
+    wchar_t *app_container_sid;
+    wchar_t security_descriptor_str[PATH_MAX];
+    int ret;
+
+    /* Resolve user's SID for compatibility with wine */
+    user_sid = _pyi_win32_get_sid(TokenUser);
+
+    /* If program is running within an AppContainer, the app container SID has to be added to
+     * the DACL, otherwise our process will not have access to the temporary directory. */
+    app_container_sid = _pyi_win32_get_sid(TokenAppContainerSid); /* NULL when not running in AppContainer */
+
+    /* DACL descriptor D:dacl_flags(string_ace1)(string_ace2)
+     * with ACE string:
+     * ace_type;ace_flags;rights;object_guid;inherit_object_guid;account_sid;(resource_attribute)
+     * - ace_type = SDDL_ACCESS_ALLOWED (A)
+     * - rights = SDDL_FILE_ALL (FA)
+     * - account_sid = current user (queried SID)
+     */
+    if (app_container_sid) {
+        ret = _snwprintf(
+            security_descriptor_str,
+            PATH_MAX,
+            L"D:(A;;FA;;;%s)(A;;FA;;;%s)",
+            user_sid ? user_sid : L"S-1-3-4",
+            app_container_sid);
+    } else {
+        ret = _snwprintf(
+            security_descriptor_str,
+            PATH_MAX,
+            L"D:(A;;FA;;;%s)",
+            user_sid ? user_sid : L"S-1-3-4");
+    }
+
+    LocalFree(user_sid); /* Must be freed using LocalFree() */
+    LocalFree(app_container_sid); /* Must be freed using LocalFree() */
+
+    if (ret >= PATH_MAX) {
+        OTHERERROR("Security descriptor string length exceeds PATH_MAX!\n");
+        return NULL;
+    }
+
+    /* Convert security descriptor string to security descriptor, and
+     * store it in the SECURITY_ATTRIBUTES structure. */
+    VS("LOADER: initializing security descriptor from string: %S\n", security_descriptor_str);
+    ret = ConvertStringSecurityDescriptorToSecurityDescriptorW(
+        security_descriptor_str,
+        SDDL_REVISION_1,
+        &lpSecurityDescriptor,
+        NULL);
+    if (ret == 0) {
+        return NULL;
+    }
+
+    /* Allocate SECURITY_ATTRIBUTES and fill it in. */
+    security_attr = calloc(1, sizeof(SECURITY_ATTRIBUTES));
+    security_attr->nLength = sizeof(SECURITY_ATTRIBUTES);
+    security_attr->bInheritHandle = FALSE;
+    security_attr->lpSecurityDescriptor = lpSecurityDescriptor;
+
+    return security_attr;
+}
+
+/* Free security descriptor applied to application's temporary directory and its
+ * sub-directories.
+ */
+void
+pyi_win32_free_security_descriptor(SECURITY_ATTRIBUTES **security_attr_ref)
+{
+    SECURITY_ATTRIBUTES *security_attr = *security_attr_ref;
+
+    security_attr_ref = NULL;
+
+    /* Free security descriptor */
+    LocalFree(security_attr->lpSecurityDescriptor);
+
+    /* Free the structure itself */
+    free(security_attr);
+}
+
+
+/**********************************************************************\
+ *      Console minimization/hiding (console-enabled build only)      *
+\**********************************************************************/
+#if !defined(WINDOWED)
+
+/* Helper that hides or minimizes the console window if it is owned by
+ * the process. The show_cmd argument is passed to the ShowWindow call,
+ * and should be either SW_HIDE or SW_SHOWMINNOACTIVE.
+ */
+static void pyi_win32_adjust_console(int show_cmd)
+{
+    HWND hConsole = GetConsoleWindow();
+    if (hConsole != NULL) {
+        DWORD dwProcessId = GetCurrentProcessId();
+        DWORD dwConsoleProcessId;
+
+        if (GetWindowThreadProcessId(hConsole, &dwConsoleProcessId) == 0) {
+            return; /* Window handle is invalid */
+        }
+
+        if (dwProcessId == dwConsoleProcessId) {
+            ShowWindow(hConsole, show_cmd);
+        }
+    }
+}
+
+void pyi_win32_hide_console()
+{
+    pyi_win32_adjust_console(SW_HIDE);
+}
+
+void pyi_win32_minimize_console()
+{
+    pyi_win32_adjust_console(SW_SHOWMINNOACTIVE);
+}
+
+#endif /* !defined(WINDOWED) */
 
 
 #endif /* ifdef _WIN32 */
